Added UI Options
Added the ability for the user to open a File Browser to select their file. When they do, the importer will automatically set the appropriate data (including the added Unity Scene Name and XML File Name). It will also save the File Path the User Opened Last for Ease of File Navigation.

It will also attempt to determine if the user used sub-folders and set the Base File Path, Used Sub-Folders, XML Folder and Export Folder fields appropriately. Depending on the user's set up and how they exported, this may not work as intended. The User can, if they wish, add and remove the words that the importer looks for when looking for Sub-Folders (in the KSIUI.cs script).